pyproject.toml
[tool.poetry]
name = "mwr_l12l2"
version = "0.1.0"
description = "This sets up the retrieval of humidity and temperature from ground-based microwave radiometers and is used in E-PROFILE"
authors = ["Rolf Rüfenacht <rruefenacht@users.noreply.github.com>"]
license = "BSD 3-Clause"

[tool.poetry.dependencies]
python = "^3.8"
PyYAML = "^6.0"
numpy = "^1.24.3"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
